Update Get-OneNoteTags.psm1
First off, thanks for the great tool! I was looking for a quick way to extract all Tags (To Do's) in my OneNote pages.
I have updated the section used to get Tags, as in some cases the XML read was returning either a System.Object[] (when more than 1 TagDef is retrieved) or a single XmlElement (when only one is retrieved). Because of this, the "count" method used for $countTags was failing for single-element XmlElement, and subsequently not all Tags in a Notebook/Page/Section were properly retrieved. The basic step of the change is to cast the xml.TagRef to an array before counting them.
I tested the baseline on my OneNote and noticed that not all To Do's were captured... some debugging after, now seems to be OK.
Cheers